refactor: migrate setup redistribution/naming suites to scenario tables (#4273)
Add shared scenario harness under test/setup/support and densify three
near-cap redistribution/naming suites into thin runners plus scenario
tables (slice D partial).
